chore: move types to `core`
